+++ typescript/http-registry-server/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @hyperlane-xyz/http-registry-server
 
+## 21.1.0
+
+### Patch Changes
+
+- @hyperlane-xyz/sdk@21.1.0
+- @hyperlane-xyz/utils@21.1.0
+
+## 21.0.0
+
+### Patch Changes
+
+- Updated dependencies [c08fa32]
+- Updated dependencies [68310db]
+- Updated dependencies [b6b206d]
+- Updated dependencies [bc8b22f]
+- Updated dependencies [ed10fc1]
+- Updated dependencies [0bce4e7]
+  - @hyperlane-xyz/sdk@21.0.0
+  - @hyperlane-xyz/utils@21.0.0
+
 ## 20.1.0
 
 ### Patch Changes